types

+++ grid.py
@@ -4,6 +4,7 @@ import flaschen_np
 import time
 import argparse
 
+
 def main():
     parser = argparse.ArgumentParser(description='Grid',
                                      formatter_class=lambda prog: argparse.ArgumentDefaultsHelpFormatter(prog))
@@ -13,14 +14,14 @@ def main():
     parser.add_argument('--height', type=int, default=35, help='Canvas height')
     parser.add_argument('--width', type=int, default=45, help='Canvas width')
     parser.add_argument('--layer', '-l', type=int, default=11, help='Canvas layer (0-15)')
-    parser.add_argument('--time', '-t', type=float, default=30, help='How long to run for before exiting')
+    parser.add_argument('--time', '-t', type=int, default=30, help='How long to run for before exiting')
     parser.add_argument('--sleep', '-s', type=float, default=.005, help='How long to sleep after each line')
     
     args = parser.parse_args()
     
     ff = flaschen_np.FlaschenNP(args.host, args.port, args.width, args.height, args.layer)
 
-    for t in range(int(args.time)):
+    for t in range(args.time):
         for i in list(range(255)) + list(range(255)[::-1]):
             for x in range(args.width):
                 for y in range(args.height):
